refactor app: centralise most processing logic
this way it'll be easier to add the video stream as a separately
launchable app without having a lot of duplicate code.
