+++ src/app_stored_images/app.py
@@ -1,50 +1,37 @@
 __author__ = "KLSA"
 __copyright__ = "Copyright 2023 to Infinity and Beyond, CEDES AG"
 
+from typing import Optional
+
 import cv2
-from image_proc.image_proc import ImageProcessor
-from pattern_logger.console_logger import ConsolePatternLogger
-from pattern_logger.csv_logger import CsvPatternLogger
-from pattern_processor.pattern_filter import PatternFilter
-from visualizer.visualizer import Visualizer
+
+from app import App
 
 
-class App:
+class StoredImageApp(App):
+    """
+    Reads images from the disk, runs the pattern recognition on them and displays the images with the result.
+    """
+
     def __init__(self):
-        self.image_processor = ImageProcessor()
-        self.visualizer = Visualizer("MyVis")
-        self.loggers = [
-            ConsolePatternLogger(),
-            CsvPatternLogger(),
-        ]
-        self.pattern_filter = PatternFilter()
+        self._current_image = 0
+        self._num_images = 5  # quick-win
 
-    def run(self):
-        try:
-            num_images = 5  # quick-win
-            for i in range(num_images):
-                print(f"processing image {i}")
+        super().__init__()
 
-                frame = cv2.imread(f"assets/test_image{i:02}.jpg", cv2.IMREAD_COLOR)
-                pattern_list = self.image_processor.process(frame)
+    def get_next_frame(self) -> Optional[cv2.Mat]:
+        if self._current_image >= self._num_images:
+            return None
 
-                filtered_patterns = self.pattern_filter.set_and_filter_patterns(pattern_list)
+        i = self._current_image
+        self._current_image += 1
 
-                for logger in self.loggers:
-                    logger.log(filtered_patterns)
-
-                self.visualizer.visualize(frame, pattern_list)
-
-                if cv2.waitKey() & 0xFF == ord("q"):
-                    break
-        except KeyboardInterrupt:
-            pass
-        finally:
-            cv2.destroyAllWindows()
+        print(f"processing image {i}")
+        return cv2.imread(f"assets/test_image{i:02}.jpg", cv2.IMREAD_COLOR)
 
 
 def main():
-    app = App()
+    app = StoredImageApp()
     app.run()
